added functionality to swtich language to settingsManager script and deleted old, unused LanguageSelector script
