--- Maker/Assets/Code/Localization/LanguageSelector.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 09d44cfac601c436c9f266ccb5a9dbbb
